requirements.txt
Django==4.1.9
whitenoise~=6.2.0
python-dotenv~=0.20.0
djangorestframework~=3.14.0
mediapipe~=0.9.2.1
django-jazzmin==2.6.0
django-browser-reload==1.8.0
django-rest-knox==4.2.0
sentry-sdk==1.28.1
yookassa==2.4.0